refactor(test): extract shared createBbDesktopApi stub factory
Six test files each hand-rolled a near-identical BbDesktopApi stub. Hoist
a single createBbDesktopApi(info, browser?) into bb-desktop-test-utils and
route them through it. The two stubs that drive live info changes or assert
on method spies (useUpdateAvailableToast, useDesktopThemeSync) and the
Ladle story keep their bespoke builders.
