+++ apps/app/src/test/bb-desktop-test-utils.ts
@@ -1,4 +1,8 @@
-import type { BbDesktopBrowserApi } from "@bb/server-contract";
+import type {
+  BbDesktopApi,
+  BbDesktopBrowserApi,
+  BbDesktopInfo,
+} from "@bb/server-contract";
 
 /**
  * A no-op {@link BbDesktopBrowserApi} for tests that build a full
@@ -22,5 +26,33 @@ export function createNoopDesktopBrowserApi(): BbDesktopBrowserApi {
     onOpenTab() {
       return () => {};
     },
+  };
+}
+
+/**
+ * A full {@link BbDesktopApi} stub for tests that need `window.bbDesktop`. The
+ * update/info methods echo `info`; theme and external-open are no-ops. Pass a
+ * custom `browser` to exercise the browser control surface. Tests that drive
+ * live info changes or assert on method spies build their own stub instead.
+ */
+export function createBbDesktopApi(
+  info: BbDesktopInfo,
+  browser: BbDesktopBrowserApi = createNoopDesktopBrowserApi(),
+): BbDesktopApi {
+  return {
+    ...info,
+    browser,
+    async checkForUpdates() {
+      return info;
+    },
+    async getInfo() {
+      return info;
+    },
+    async installUpdate() {},
+    onChange() {
+      return () => {};
+    },
+    setTheme() {},
+    openExternalUrl() {},
   };
 }
